requirements.txt
faker
pandas
streamlit
pyarrow
pytest
apache-airflow==2.5.0  # optional; pin if using
